refactor(header): split into smaller components
